requirements.txt
absl-py==2.1.0
asttokens==3.0.0
astunparse==1.6.3
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
Cython==3.0.7
debugpy==1.8.12
decorator==5.2.1
exceptiongroup==1.2.2
executing==2.2.0
flatbuffers==1.12
fonttools==4.56.0
gast==0.4.0
google-auth==2.38.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.70.0
h5py==3.13.0
idna==3.10
imageio==2.37.0
imgaug==0.4.0
ipykernel==6.29.5
ipython==8.33.0
jedi==0.19.2
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
keras==2.9.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.8
lazy_loader==0.4
libclang==18.1.1
Markdown==3.7
MarkupSafe==3.0.2
matplotlib==3.6.0
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
networkx==3.4.2
numpy==1.23.0
oauthlib==3.2.2
opencv-python==4.7.0.72
opt_einsum==3.4.0
packaging==24.2
pandas==1.4.0
parso==0.8.4
pillow==11.1.0
pip==23.0.1
platformdirs==4.3.6
prompt_toolkit==3.0.50
protobuf==3.19.6
psutil==7.0.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
Pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2025.1
PyWavelets==1.8.0
pywin32==308
pyzmq==26.2.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
scikit-image==0.20.0
scikit-learn==1.1.0
scipy==1.10.0
setuptools==65.5.0
shapely==2.0.7
six==1.17.0
stack-data==0.6.3
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.2
tensorflow-estimator==2.9.0
tensorflow-gpu==2.9.2
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.5.0
threadpoolctl==3.5.0
tifffile==2025.2.18
tornado==6.4.2
tqdm==4.64.1
traitlets==5.14.3
typing_extensions==4.12.2
urllib3==2.3.0
wcwidth==0.2.13
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
